backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
deprecation==2.1.0
Django==5.2.1
Farama-Notifications==0.0.4
filelock==3.18.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.0
gast==0.6.0
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gotrue==2.12.3
grpcio==1.73.1
gymnasium==1.1.1
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.0
hyperframe==6.1.0
idna==3.10
imbalanced-learn==0.9.0
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mypy_extensions==1.1.0
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.2
pathspec==0.12.1
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.8
postgrest==1.1.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.5.3
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-learn==1.5.0
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
